Introduction of Global Unit Settings. And some partial classes made to sort some functionality

+++ Assets/Scripts/UnitStatistics.cs
@@ -10,7 +10,7 @@ namespace CM.Units
         public Statistic<int> defense;
         
         public Statistic<float> attackSpeed;
-        public Statistic<int> vision;
+        public Statistic<float> vision;
 
         public Statistic<int> strength;
         public Statistic<int> agility;
